Factored out user category menus and IO utility methods
