+++ src/test/BrokerMenuTest.java
@@ -8,14 +8,14 @@ import org.junit.Test;
 
 import main.BankManager;
 import main.BrokerMenu;
-import main.MainMenu;
+import main.IOUtils;
 import main.Security;
 
 public class BrokerMenuTest {
 
     public BrokerMenu setUpBroker() {
         BankManager bank = new BankManager();
-        BrokerMenu broker = new BrokerMenu(new MainMenu(), bank);
+        BrokerMenu broker = new BrokerMenu(new IOUtils(bank), bank);
         bank.createAccount("test");
         bank.switchAccounts(1);
         ArrayList<Security> portfolio = bank.getPortfolio();
